seller-cabinet/requirements.txt
django==5.2.12
djangorestframework==3.17.1
psycopg==3.3.3
pyjwt==2.12.1
python-dotenv==1.2.2
